python/swap-and-pay/requirements.txt
axonfi>=0.3.2
python-dotenv>=1.0.0
